my-app-backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
python-multipart
pydantic
google-generativeai
python-dotenv
greenlet
PyJWT
passlib[bcrypt]